requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
deprecation==2.1.0
Django==5.0.7
djangorestframework==3.15.2
gotrue==2.6.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
packaging==24.1
postgrest==0.16.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
realtime==1.0.6
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
typing_extensions==4.12.2
websockets==12.0
